Advanced College Projects
The most advanced thing I have on file is the PHP-SQL website I made in an advanced Web Programming class. It was pretty easy, to be honest, mostly just figuring out compatibility issues between PHP and SQL, because of the change to "sqli" functions. There 2 programs, Object Sorting which was harder than I though it would be! Also, password strength introduced me to input validation and optimization.